requirements.txt
streamlit>=1.33.0
llama-index-core>=0.10.34
llama-index-llms-ollama>=0.1.6
llama-index-embeddings-huggingface>=0.1.4
python-dotenv>=1.0.0